feat: refact code
