+++ packages/indexer/tests/indexer.test.js
@@ -2,7 +2,7 @@ const test = require("ava");
 const sinon = require("sinon");
 const fs = require("fs");
 const { Indexer } = require("../lib");
-const { indexerSubscribeTestCases } = require("./indexer_subscribe_cases.js");
+const { lock, type, indexerSubscribeTestCases } = require("./test_cases.js");
 // the node_uri will not be connected during the test process, only serves as a placeholder when create an indexer instance.
 const node_uri = "http://127.0.0.1:8116";
 const tmpIndexedDataPath = __dirname + "/tmp_indexed_data3";
@@ -29,4 +29,29 @@ test("test indexer subscribe by differen queryOptions", async (t) => {
     // clear rocksdb test data
     fs.rmdirSync(tmpIndexedDataPath, { recursive: true });
   }
+});
+
+test("throw error when pass both lock and type to subscribe", (t) => {
+  const error = t.throws(
+    () => {
+      const queryOption = { lock: lock, type: type };
+      indexer.subscribe(queryOption);
+    },
+    { instanceOf: Error }
+  );
+  t.is(
+    error.message,
+    "The notification machanism only supports you subscribing for one script once so far!"
+  );
+});
+
+test("throw error when pass both null lock and null type to subscribe", (t) => {
+  const error = t.throws(
+    () => {
+      const queryOption = {};
+      indexer.subscribe(queryOption);
+    },
+    { instanceOf: Error }
+  );
+  t.is(error.message, "Either lock or type script must be provided!");
 });
